requirements.txt
imageio==2.31.6
matplotlib>=3.8.0
numpy>=1.26.1
opencv-python>=4.8.1.78
Pillow>=9.5.0
pillow-avif-plugin
pillow-heif
scikit-image>=0.20.0
scikit-learn>=1.3.2
ttkthemes>=3.2.2
send2trash
openai-clip
torch
ttkthemes
python-vlc
